Moved DB related classes to their own namespace
