+++ src/ContentRelationshipMigrator.php
@@ -6,6 +6,7 @@ namespace Inpsyde\MultilingualPress2to3;
 use Dhii\I18n\FormatTranslatorInterface;
 use Dhii\I18n\StringTranslatingTrait;
 use Dhii\I18n\StringTranslatorAwareTrait;
+use Inpsyde\MultilingualPress2to3\Db\DatabaseWpdbTrait;
 use Throwable;
 use UnexpectedValueException;
 use wpdb as Wpdb;
