+++ src/MigrateRelationshipsCliCommand.php
@@ -7,6 +7,7 @@ use Dhii\I18n\FormatTranslatorInterface;
 use Dhii\I18n\StringTranslatorConsumingTrait;
 use Exception;
 use Inpsyde\MultilingualPress2to3\Cli\CliUtilsTrait;
+use Inpsyde\MultilingualPress2to3\Db\DatabaseWpdbTrait;
 use Throwable;
 use wpdb as Wpdb;
